+++ src/components/Chat/ChatMessage.tsx
@@ -132,7 +132,7 @@ class ChatMessage extends UIComponent<Extendable<IChatMessageProps>, any> {
 
     const authorComponent = Text.create(author, {
       defaultProps: {
-        size: 'sm',
+        size: 'small',
         styles: styles.author,
         variables: variables.author,
       },
@@ -140,7 +140,7 @@ class ChatMessage extends UIComponent<Extendable<IChatMessageProps>, any> {
 
     const timestampComponent = Text.create(timestamp, {
       defaultProps: {
-        size: 'sm',
+        size: 'small',
         timestamp: true,
         styles: styles.timestamp,
         variables: variables.timestamp,