+++ src/lib/accessibility/Behaviors/Popup/PopupBehavior.ts
@@ -19,7 +19,7 @@ const PopupBehavior: Accessibility = (props: any) => ({
   },
   keyActions: {
     trigger: {
-      performClick: {
+      toggle: {
         keyCombinations: [{ keyCode: keyboardKey.Enter }, { keyCode: keyboardKey.Spacebar }],
       },
     },