+++ docs/src/routes.tsx
@@ -11,6 +11,7 @@ import QuickStart from './views/QuickStart'
 import Accessibility from './views/Accessibility'
 import Theming from './views/Theming'
 import ThemingExamples from './views/ThemingExamples'
+import Glossary from './views/Glossary'
 import { Provider, themes } from '@stardust-ui/react'
 
 import { mergeThemes } from '../../src/lib'
@@ -47,6 +48,7 @@ class Router extends React.Component<any, any> {
               <DocsLayout exact path="/" component={Introduction} />
               <DocsLayout exact path="/:type/:name" component={DocsRoot} sidebar />
               <DocsLayout exact path="/quick-start" component={QuickStart} />
+              <DocsLayout exact path="/glossary" component={Glossary} />
               <DocsLayout exact path="/accessibility" component={Accessibility} />
               <DocsLayout exact path="/theming" component={Theming} />
               <DocsLayout exact path="/theming-examples" component={ThemingExamples} />