+++ docs/src/examples/components/Popup/Types/PopupExample.tsx
@@ -8,17 +8,20 @@ class PopupExample extends React.Component<any, any> {
     this.setState(prev => ({ popupOpen: !prev.popupOpen }))
   }
 
+  handleOpenChange = (e, newProps) => {
+    this.setState({ popupOpen: newProps.open })
+    alert(`Popup open state was changed to "${newProps.open ? 'true' : 'false'}".`)
+  }
+
   render() {
     return (
       <Popup
         open={this.state.popupOpen}
-        onOpenChange={(e, newProps) => {
-          this.setState({ popupOpen: newProps.open })
-          alert(`Popup open state was changed to "${newProps.open ? 'true' : 'false'}".`)
-        }}
-        trigger={<Button icon="expand" onClick={() => this.togglePopup()} />}
+        onOpenChange={this.handleOpenChange.bind(this)}
         content="Add users to your feed."
-      />
+      >
+        <Button icon="expand" onClick={() => this.togglePopup()} />
+      </Popup>
     )
   }
 }